Thursday's commit
Worked on the Magic Circle today, made it so Character now moves into the middle of the circle when it enters the magic circle's collider box

Will probably work on layer fixing soon, so the Character doesn't walk on palms, but rather behind them. Will probably have to remove the palms from the tilemap and add them as objects, then give them colliderboxes. Needs researching.