Monday 16th dec
Did some more work in Character Creation, fixing some code. Getting closer to a functioning solution.
Made a new method to put attributes and racials together in stead of calling it from checkrace(). Working on a fix for all the AddToAttribute methods, current way of doing it looks promising, will finish it tomorrow and see if I can do the same for SubstractFromAttribute methods.