Thursday 19th dec
Tried to follow a tutorial on serializing. Not working as intended yet, need to get back to the drawing board and look at the things from a distance. See what's going wrong.  Otherwise save and load to/from binary file seem to work. Just need it to work.... As intended.